# openapi.yml
openapi: '3.0.2'
info:
  title: simple-twitter
  version: '1.0'
  description: '提供貼文、留言和按讚服務的內容分享平台'
servers:
  - description: '開發伺服器'
    url: http://localhost:3000/api
...
paths:
...
  /myInfo:
    get:
      tags:
        - general
      operationId: getMyInfo
      summary: 當前使用者資訊
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功獲取個人資訊
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/myInfo'
  # tweets
...
  /users/{user_id}:
    get:
      tags:
        - users
      operationId: getUserById
      summary: 指定帳號資訊
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
            default: 20
          description: 帳號 id
      responses:
        '200':
          description: 成功獲取帳號資訊
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fullUserInfo'
        '400':
          description: 沒有該id的帳號
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/noAccountError'
    put:
      tags:
        - users
      operationId: updateUserById
      summary: 更新帳號資訊
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
            default: 20
          description: 自身帳號 id
      requestBody:
        description: 修改帳號的請求
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/myInfo'
            examples:
              editNameAndIntro:
                summary: 更新名稱 & 自我介紹 & 大頭貼 & 背景圖
                value:
                  name: new name
                  introduction: new intro
                  avatar: new avatar
                  cover: new cover
              editAll:
                summary: 更新帳號 & 名稱 & 信箱，並輸入舊密碼
                value:
                  account: new account
                  name: new name
                  email: new email
                  password: old password
                  checkPassword: old password

      responses:
        '200':
          description: 成功修改帳號
          content:
            application/json:
              examples:
                notEditAccount:
                  summary: 未修改account欄位的值
                editAccount:
                  summary: 修改account欄位的值
                  value:
                    accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2NvdW50IjoidXNlcjEiLCJ0eXBlIjoiYWNjZXNzIiwiaWF0IjoxNjg1OTUxMTgwLCJleHAiOjE2ODU5NTEyMTB9.7CfezEjXoCmBwzqwXNdI7Q6a39xfeRlpgssBPZswV8g

        '400':
          description: 更新失敗
          content:
            application/json:
              examples:
                editSelfError:
                  summary: 帳號id不是自己
                  value:
                    message: 只能修改自己的帳號資訊
                checkPasswordError:
                  summary: 確認密碼和密碼不一致
                  value:
                    message: 密碼和確認密碼不一致
                passwordError:
                  summary: 密碼驗證錯誤
                  value:
                    message: 密碼有誤
                dupAccountError:
                  summary: 帳號已經被使用過了
                  value:
                    message: 帳號已經被使用過了
                dupNameError:
                  summary: 名稱已經被使用過了
                  value:
                    message: 名稱已經被使用過了
        '507':
          description: 資料庫請求錯誤
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dbError'
  /users/{user_id}/followers:
    get:
      tags:
        - users
      operationId: getUserFollowers
      summary: 指定帳號的粉絲名單
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: integer
            example: 2
          required: true
          description: 帳號 id
      responses:
        '200':
          description: 成功獲取該帳號的粉絲名單
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: '#/components/schemas/userInfoWithIntro'
                    - $ref: '#/components/schemas/userInfoWithIntro'
        '400':
          description: 帳號不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/noAccountError'
  /users/{user_id}/followings:
    get:
      tags:
        - users
      operationId: getUserFollowings
      summary: 指定帳號的追蹤中名單
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: integer
            example: 2
          required: true
          description: 帳號 id
      responses:
        '200':
          description: 成功獲取該帳號的追蹤中名單
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: '#/components/schemas/userInfoWithIntro'
                    - $ref: '#/components/schemas/userInfoWithIntro'
        '400':
          description: 帳號不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/noAccountError'
  /users/{user_id}/likes:
    get:
      tags:
        - users
      operationId: getUserLikes
      summary: 指定帳號的按讚貼文
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: integer
            example: 2
          required: true
          description: 帳號 id
      responses:
        '200':
          description: 成功獲取該帳號喜歡的貼文
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: '#/components/schemas/tweet'
                    - $ref: '#/components/schemas/tweet'
        '400':
          description: 帳號不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/noAccountError'
  /users/{user_id}/replied_tweets:
    get:
      tags:
        - users
      operationId: getUserTweetsReply
      summary: 指定帳號回覆過的貼文
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: integer
            example: 1
          required: true
          description: 帳號 id
      responses:
        '200':
          description: 成功獲取該帳號回覆過的貼文
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: '#/components/schemas/replyWithTweetUser'
                    - $ref: '#/components/schemas/replyWithTweetUser'
        '400':
          description: 帳號不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/noAccountError'
  /users/{user_id}/tweets:
    get:
      tags:
        - users
      operationId: getUserTweets
      summary: 指定帳號的貼文
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: user_id
          schema:
            type: integer
            example: 1
          required: true
          description: 帳號 id
      responses:
        '200':
          description: 成功獲取該帳號的貼文
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: '#/components/schemas/tweet'
                    - $ref: '#/components/schemas/tweet'
        '400':
          description: 帳號不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/noAccountError'
  # followships
...
  /admin/users:
    get:
      tags:
        - admin
      operationId: getAllUser
      summary: 獲取所有帳號資訊
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功獲取所有使用者資訊
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: '#/components/schemas/adminUserInfo'
                    - $ref: '#/components/schemas/adminUserInfo'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'access token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2NvdW50IjoidXNlcjEiLCJ0eXBlIjoiYWNjZXNzIn0.j2QXHeg2HFQNMcmmTeuZjrjSxPe8whGsqrMBjGRz0xg'
  schemas:
...
    myInfo:
      properties:
        id:
          type: integer
          example: 1
        account:
          type: string
          example: user1
        name:
          type: string
          example: user1
        email:
          type: string
          example: user1@example.com
        role:
          type: string
          example: user
        avatar:
          type: string
          example: null
...
    tweet:
      properties:
        id:
          type: integer
          example: 2
        description:
          type: string
          example: this is description of tweet
        createdAt:
          type: string
          example: '2023-05-31T03:00:07.000Z'
        updatedAt:
          type: string
          example: '2023-05-31T03:00:07.000Z'
        replyCount:
          type: integer
          example: 1
        likeCount:
          type: integer
          example: 1
        liked:
          type: boolean
          example: 1
        User:
          type: object
          $ref: '#/components/schemas/userInfo'
...
    userInfo:
      properties:
        id:
          type: integer
          example: 20
        account:
          type: string
          example: user1
        name:
          type: string
          example: 12345678
        avatar:
          type: string
          example: null
    userInfoWithIntro:
      allOf:
        - $ref: '#/components/schemas/userInfo'
      properties:
        introduction:
          type: string
          example: this is intro
    reply:
      properties:
        id:
          type: integer
          example: 3
        TweetId:
          type: integer
          example: 1
        comment:
          type: string
          example: this is test comment
        createdAt:
          type: string
          example: '2023-06-01T13:30:00.000Z'
        reply_user:
          type: object
          $ref: '#/components/schemas/userInfo'
    replyWithTweetUser:
      allOf:
        - $ref: '#/components/schemas/reply'
      properties:
        tweet_user:
          type: object
          $ref: '#/components/schemas/userInfo'
    fullUserInfo:
      allOf:
        - $ref: '#/components/schemas/userInfo'
      properties:
        email:
          type: string
          example: user1@example.com
        role:
          type: string
          example: 'user'
        introduction:
          type: string
          example: introduction of user
        followed:
          type: boolean
          example: 1
        cover:
          type: string
          example: null
    adminUserInfo:
      allOf:
        - $ref: '#/components/schemas/userInfo'
      properties:
        cover:
          type: string
          example: null
        followed:
          type: boolean
          example: 1
        tweet_count:
          type: number
          example: 10
        like_count:
          type: number
          example: 1
        following_count:
          type: number
          example: 0
        follower_count:
          type: number
          example: 1
...
    noAccountError:
      properties:
        message:
          type: string
          example: 帳號不存在
...
    dbError:
      properties:
        message:
          type: string
          example: 資料庫請求錯誤